pyproject.toml
[project]
name = "fml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.2",
    "jaxtyping>=0.2.36",
    "matplotlib>=3.9.2",
    "numpy>=2.1.3",
    "pillow-simd>=9.5.0.post2",
    "setuptools>=75.6.0",
    "tensorboard>=2.18.0",
    "timm>=1.0.11",
    "torch>=2.5.1",
    "triton>=3.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.pytest.ini_options]
testpaths = ["tests", "integration"]
pythonpath = ["."]

[dependency-groups]
dev = [
    "hatchling>=1.26.3",
    "ipython>=8.29.0",
    "matplotlib>=3.9.2",
    "pytest>=8.3.3",
    "setuptools>=75.6.0",
]
train = [
    "hydra-core>=1.3.2",
    "matplotlib>=3.9.2",
    "tensorboard>=2.18.0",
    "triton>=3.1.0",
]
